Begin virtualizing chats
