+++ src/modules/aifn/react/react.ts
@@ -1,7 +1,7 @@
 import { DLLMId } from '~/modules/llms/llm.types';
 import { OpenAI } from '~/modules/openai/openai.types';
 import { callApiSearchGoogle } from '~/modules/search/search.client';
-import { callChat } from '~/modules/openai/openai.client';
+import { callChat } from '~/modules/llms/llm.client';
 
 import { currentDate, reActPrompt } from '../prompts';
