+++ src/modules/llms/openai/openai.client.ts
@@ -1,0 +1,38 @@
+import { DLLM } from '~/modules/llms/llm.types';
+import { OpenAI } from '~/modules/openai/openai.types';
+import { apiAsync } from '~/modules/trpc/trpc.client';
+
+import { normalizeSetup, SourceSetupOpenAI } from './vendor';
+
+
+export const hasServerKeyOpenAI = !!process.env.HAS_SERVER_KEY_OPENAI;
+
+export const isValidOpenAIApiKey = (apiKey?: string) => !!apiKey && apiKey.startsWith('sk-') && apiKey.length > 40;
+
+
+/**
+ * This function either returns the LLM response, or throws a descriptive error string
+ */
+export async function callChat(llm: DLLM, messages: OpenAI.Wire.Chat.Message[], maxTokens?: number): Promise<OpenAI.API.Chat.Response> {
+  // access params (source)
+  const partialSetup = llm._source.setup as Partial<SourceSetupOpenAI>;
+  const sourceSetupOpenAI = normalizeSetup(partialSetup);
+
+  // model params (llm)
+  const openaiLlmRef = llm.options.llmRef!;
+  const modelTemperature = llm.options.llmTemperature || 0.5;
+  // const maxTokens = llm.options.llmResponseTokens || 1024; // <- note: this would be for chat answers, not programmatic chat calls
+
+  try {
+    return await apiAsync.openai.chatGenerate.mutate({
+      access: sourceSetupOpenAI,
+      history: messages,
+      model: { id: openaiLlmRef, temperature: modelTemperature, ...(maxTokens && { maxTokens }) },
+    });
+    // errorMessage = `issue fetching: ${response.status} · ${response.statusText}${errorPayload ? ' · ' + JSON.stringify(errorPayload) : ''}`;
+  } catch (error: any) {
+    const errorMessage = `fetch error: ${error?.message || error?.toString() || 'Unknown error'}`;
+    console.error(`callChat: ${errorMessage}`);
+    throw new Error(errorMessage);
+  }
+}
